Add some function description and other methods

+++ code/Modules/utilities.py
@@ -2,7 +2,7 @@
 import matplotlib.pyplot as plt
 import seaborn as sn
 import os
-from pandas import read_csv
+from pandas import read_csv, read_pickle
 from Modules.config import *
 
 
@@ -17,11 +17,29 @@ def plot_dataframe(data, target_column=None):
     plt.show()
 
 
-def csv2pickle(filename):
+def count_missing_values(dataframe):
+    missing_values = dataframe.isnull().sum()
+    print(missing_values)
+    return missing_values.sum() == 0
+
+
+def csv2pickle(filename, remove=True):
     path = os.path.join(base_dir, filename)
     df = read_csv(path, sep=',')
-    os.remove(path)
+    if remove:
+        os.remove(path)
     filename = filename.split('.')[0] + '.pkl'
     dataframe_path = os.path.join(base_dir, filename)
     df.to_pickle(dataframe_path)
+    return dataframe_path
+
+
+def pickle2csv(filename, remove=False):
+    path = os.path.join(base_dir, filename)
+    df = read_pickle(path)
+    if remove:
+        os.remove(path)
+    filename = filename.split('.')[0] + '.csv'
+    dataframe_path = os.path.join(base_dir, filename)
+    df.to_csv(dataframe_path)
     return dataframe_path
